reduced redundancy in tsdb, an attempt for universal db
